requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
asgiref==3.2.10
attrs==20.2.0
bitstring==3.1.7
cffi==1.14.2
click==7.1.2
Django==3.1.1
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
Flask-Datepicker==0.14
Flask-FontAwesome==0.1.5
Flask-RESTful==0.3.8
flask-restx==0.2.0
Flask-SocketIO==4.3.1
Flask-WTF==0.14.3
future==0.18.2
gevent==20.6.2
greenlet==0.4.16
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lxml==4.5.2
MarkupSafe==1.1.1
ntplib==0.3.4
numpy==1.19.1
py==1.9.0
pycparser==2.20
pyrsistent==0.17.3
pyserial==3.4
pyshark==0.4.2.11
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.4
zope.interface==5.1.0
